Simplify the jscad/martinez glue code
