+++ packages/modeling/src/operations/booleans/intersectGeom2.js
@@ -1,6 +1,7 @@
 import flatten from '../../utils/flatten.js'
 
-import * as martinez from './martinez/index.js'
+import { INTERSECTION } from './martinez/operation.js'
+import boolean from './martinez/index.js'
 
 /*
  * Return a new 2D geometry representing space in both the first geometry and
@@ -13,7 +14,7 @@ export const intersectGeom2 = (...geometries) => {
 
   let newgeometry = geometries.shift()
   geometries.forEach((geometry) => {
-    newgeometry = martinez.intersection(newgeometry, geometry)
+    newgeometry = boolean(newgeometry, geometry, INTERSECTION)
   })
 
   return newgeometry
