--- packages/modeling/src/operations/booleans/martinez/martinez.js
@@ -1,104 +1,0 @@
-/*
- * Implementation of the Martinez 2D polygon clipping algorithm
- * Copyright (c) 2018 Alexander Milevski
- * https://github.com/w8r/martinez
- */
-
-import subdivideSegments from './subdivide_segments.js'
-import connectEdges from './connect_edges.js'
-import fillQueue from './fill_queue.js'
-import {
-  INTERSECTION,
-  DIFFERENCE,
-  UNION,
-  XOR
-} from './operation.js'
-
-const EMPTY = []
-
-const trivialOperation = (subject, clipping, operation) => {
-  let result = null
-  if (subject.length * clipping.length === 0) {
-    if (operation === INTERSECTION) {
-      result = EMPTY
-    } else if (operation === DIFFERENCE) {
-      result = subject
-    } else if (operation === UNION ||
-               operation === XOR) {
-      result = (subject.length === 0) ? clipping : subject
-    }
-  }
-  return result
-}
-
-const compareBBoxes = (subject, clipping, sbbox, cbbox, operation) => {
-  let result = null
-  if (sbbox[0] > cbbox[2] ||
-      cbbox[0] > sbbox[2] ||
-      sbbox[1] > cbbox[3] ||
-      cbbox[1] > sbbox[3]) {
-    if (operation === INTERSECTION) {
-      result = EMPTY
-    } else if (operation === DIFFERENCE) {
-      result = subject
-    } else if (operation === UNION ||
-               operation === XOR) {
-      result = subject.concat(clipping)
-    }
-  }
-  return result
-}
-
-export default function boolean (subject, clipping, operation) {
-  if (typeof subject[0][0][0] === 'number') {
-    subject = [subject]
-  }
-  if (typeof clipping[0][0][0] === 'number') {
-    clipping = [clipping]
-  }
-  let trivial = trivialOperation(subject, clipping, operation)
-  if (trivial) {
-    return trivial === EMPTY ? null : trivial
-  }
-  const sbbox = [Infinity, Infinity, -Infinity, -Infinity]
-  const cbbox = [Infinity, Infinity, -Infinity, -Infinity]
-
-  // console.time('fill queue')
-  const eventQueue = fillQueue(subject, clipping, sbbox, cbbox, operation)
-  // console.timeEnd('fill queue')
-
-  trivial = compareBBoxes(subject, clipping, sbbox, cbbox, operation)
-  if (trivial) {
-    return trivial === EMPTY ? null : trivial
-  }
-  // console.time('subdivide edges')
-  const sortedEvents = subdivideSegments(eventQueue, subject, clipping, sbbox, cbbox, operation)
-  // console.timeEnd('subdivide edges')
-
-  // console.time('connect vertices')
-  const contours = connectEdges(sortedEvents, operation)
-  // console.timeEnd('connect vertices')
-
-  // Convert contours to geom2
-  const polygons = []
-  for (let i = 0; i < contours.length; i++) {
-    const contour = contours[i]
-    if (contour.isExterior()) {
-      // The exterior ring goes first
-      const rings = [contour.points]
-      // Followed by holes if any
-      for (let j = 0; j < contour.holeIds.length; j++) {
-        const holeId = contour.holeIds[j]
-        const holePoints = contours[holeId].points
-        const hole = []
-        for (let k = holePoints.length - 2; k >= 0; k--) {
-          hole.push(holePoints[k])
-        }
-        rings.push(hole)
-      }
-      polygons.push(rings)
-    }
-  }
-
-  return polygons
-}
